requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
whitenoise==6.9.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
python-dotenv==1.0.1
